data_scrapper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.0.0
python-dotenv==1.0.0
tiktoken==0.5.1
